shell: Fix relational operator to check if canvas size is smaller than minimum size in pick interface

+++ shell/pick.c
@@ -83,7 +83,7 @@ static void pick_shellgrab_touchpoint_up(struct touchpoint_grab *base, uint32_t 
 			if (shell->is_logging_grab != 0)
 				nemolog_message("PICK", "[UP:SCALE] %llu: sx(%f) sy(%f) width(%d) height(%d) (%u)\n", touchid, bin->view->geometry.sx, bin->view->geometry.sy, width, height, time);
 
-			if (bin->min_width >= width || bin->min_height >= height) {
+			if (bin->min_width >= width && bin->min_height >= height) {
 				if (nemoview_has_state(bin->view, NEMOVIEW_CLOSE_STATE) == 0) {
 					if (bin->type == NEMOSHELL_SURFACE_XWAYLAND_TYPE)
 						kill(-bin->pid, SIGKILL);
@@ -406,7 +406,7 @@ static void pick_actorgrab_touchpoint_up(struct touchpoint_grab *base, uint32_t 
 			int32_t width = actor->view->content->width * actor->view->geometry.sx;
 			int32_t height = actor->view->content->height * actor->view->geometry.sy;
 
-			if (actor->min_width >= width || actor->min_height >= height) {
+			if (actor->min_width >= width && actor->min_height >= height) {
 				nemoactor_dispatch_destroy(actor);
 			} else if (nemoview_has_state(actor->view, NEMOVIEW_RESIZE_STATE) != 0) {
 				int32_t sx, sy;
